WIP Travis support

+++ bin/install-puppet.py
@@ -151,8 +151,11 @@ def install_puppet(sysname,osname=None,osver=None,osvername=None):
         gemfile = tempfile.NamedTemporaryFile(mode='w',delete=False)
         gemfile.write( "gem 'puppet'\n")
         gemfile.close()
+        sysdo(['cat',gemfile.name)
         sysdo(['bundle','install','--gemfile='+gemfile.name])
-        os.remove(gemfile.name+'.lock')
+        print 'puppet gem installed.'
+        if os.path.isfile(gemfile.name+'.lock'):
+              os.remove(gemfile.name+'.lock')
         os.remove(gemfile.name)
 
 def main():
